# Update cleanupOnDestroy_smart?