--- a/src/Utility/cleanupOnDestroy_smart.lua
+++ b/src/Utility/cleanupOnDestroy_smart.lua
@@ -18,89 +18,35 @@ local cleanup = require(Package.Utility.cleanup)
 -- The event to use for waiting (typically Heartbeat)
 local STEP_EVENT = RunService.Heartbeat
 
-local function noOp()
-	-- intentionally blank - no operation!
-end
-
-local function cleanupOnDestroy(instance: Instance?, cleanupTask: cleanup.Task): (() -> ())
+local function cleanupOnDestroy(instance: Instance?, task: cleanup.Task): (() -> ())
 	-- set up manual disconnection logic
 	local isDisconnected = false
-	local ancestryChangedConn, highTierConn
-	local ignoreHighTierConn = true
+	local ancestryChangedConn
 
 	local function disconnect()
 		if not isDisconnected then
 			isDisconnected = true
 			ancestryChangedConn:Disconnect()
-
-			if highTierConn and highTierConn.Connected ~= nil then
-				highTierConn:Disconnect()
-			end
 		end
 	end
 
 	-- We can't keep a reference to the instance, but we need to keep track of
 	-- when the instance is parented to `nil`.
 	-- To get around this, we can save the parent from AncestryChanged here
-	local isNilParented = (instance :: Instance).Parent == nil
-	local parentTreeSize = 0
-
-	local function calculateFromTree(child: Instance?)
-		parentTreeSize = 0
-
-		local finished = child == nil or child == game
-
-		if finished then
-			ignoreHighTierConn = true
-		else
-			while finished == false do
-				parentTreeSize += 1
-				local parent: Instance? = child and child.Parent or nil
-	
-				if parent == nil or parent == game then
-					if child then
-						if highTierConn then
-							highTierConn:Disconnect()
-						end
-	
-						highTierConn = child:GetPropertyChangedSignal("ClassName"):Connect(noOp)
-						ignoreHighTierConn = false
-	
-					else
-						ignoreHighTierConn = true
-					end
-					
-					finished = true
-				end
-	
-				child = parent
-			end
-		end
-		
-		isNilParented = child == nil
-		child = nil
-	end
-	
-	calculateFromTree(instance :: Instance)
+	local isNilParented = if instance then not game:IsAncestorOf(instance :: Instance) else false
 
 	-- when AncestryChanged is called, run some destroy-checking logic
 	-- this function can yield when called, so make sure to call in a new thread
 	-- if you don't want your current thread to block
 	local function onInstanceMove(_doNotUse: Instance?, newParent: Instance?)
-		local oldParentTreeSize = parentTreeSize
-
-		-- discard the first argument so we don't inhibit GC
 		if isDisconnected then
-			_doNotUse = nil
 			return
-			
-		else
-			-- determine how many objects there are until we reach `nil` or `game`
-			calculateFromTree(_doNotUse)
-			_doNotUse = nil
 		end
 
-		--isNilParented = newParent == nil
+		-- discard the first argument so we don't inhibit GC
+		_doNotUse = nil
+
+		isNilParented = not game:IsAncestorOf(newParent)
 
 		-- if the instance has been moved into a nil parent, it could possibly
 		-- have been destroyed if no other references exist
@@ -118,7 +64,7 @@ local function cleanupOnDestroy(instance: Instance?, cleanupTask: cleanup.Task):
 
 				elseif not ancestryChangedConn.Connected then
 					-- if our event was disconnected, the instance was destroyed
-					cleanup(cleanupTask)
+					cleanup(task)
 					disconnect()
 
 				else
@@ -132,39 +78,33 @@ local function cleanupOnDestroy(instance: Instance?, cleanupTask: cleanup.Task):
 					-- out of event-based options.
 					while
 						isNilParented and
-						not isDisconnected and
-						(highTierConn.Connected and ignoreHighTierConn == false) and
-						ancestryChangedConn.Connected
+						ancestryChangedConn.Connected and
+						not isDisconnected
 					do
-
 						-- FUTURE: is this too often?
 						STEP_EVENT:Wait()
 					end
-					
+
 					-- The instance was either destroyed, or we stopped looping
 					-- for another reason (reparented or `disconnect` called)
 					-- Check those other conditions before calling the callback.
-					if
-						isDisconnected or
-						not isNilParented or
-						(ancestryChangedConn.Connected and (ignoreHighTierConn == true or (highTierConn.Connected and ignoreHighTierConn == false)))
-					then
+					if isDisconnected or not isNilParented then
 						return
 					end
 
-					cleanup(cleanupTask)
+					cleanup(task)
 					disconnect()
 				end
 			end)()
 		end
 	end
 
-	ancestryChangedConn = (instance :: Instance):GetPropertyChangedSignal("Parent"):Connect(onInstanceMove)
+	ancestryChangedConn = (instance :: Instance).AncestryChanged:Connect(onInstanceMove)
 
 	-- in case the instance is currently in nil, we should call `onInstanceMove`
 	-- before any other code has the opportunity to run
 	if isNilParented then
-		onInstanceMove(instance :: Instance, (instance :: Instance).Parent)
+		onInstanceMove(nil, (instance :: Instance).Parent)
 	end
 
 	-- remove this functions' reference to the instance, so it doesn't influence