--- a/test/Utility/cleanupOnDestroy.spec.lua
+++ b/test/Utility/cleanupOnDestroy.spec.lua
@@ -1,7 +1,7 @@
 local RunService = game:GetService("RunService")
 
 local Package = game:GetService("ReplicatedStorage").Fusion
-local cleanupOnDestroy = require(Package.Utility.cleanupOnDestroy)
+local cleanupOnDestroy = require(Package.Utility.cleanupOnDestroy_smart)
 
 local function waitForDefer()
 	RunService.RenderStepped:Wait()
@@ -111,7 +111,7 @@ return function()
 		ins.Parent = nil
 
 		local start = os.clock()
-		local timeout = 3
+		local timeout = 6
 
 		repeat
 			RunService.RenderStepped:Wait()
@@ -121,5 +121,96 @@ return function()
 		until done
 
 		error("Instance was incorrectly collected")
+	end)
+
+	it("should not run tasks while reference not held but parent reference is held in nil", function()
+		local parent = Instance.new("Folder")
+		parent.Parent = workspace
+
+		local done = false
+		local function callback()
+			done = true
+		end
+
+		do
+			local ins = Instance.new("Folder")
+			ins.Parent = parent
+			cleanupOnDestroy(ins, callback)
+		end
+
+		parent.Parent = nil
+
+		local start = os.clock()
+		local timeout = 6
+
+		repeat
+			RunService.RenderStepped:Wait()
+			if os.clock() - start > timeout then
+				return
+			end
+		until done
+
+		error("Instance was incorrectly collected")
+	end)
+
+	it("should not run tasks when parent is in nil, gets removed from parent and parent gets destroyed", function()
+		local parent = Instance.new("Folder")
+		local ins = Instance.new("Folder")
+		ins.Parent = parent
+
+		local done = false
+		local function callback()
+			done = true
+		end
+
+		cleanupOnDestroy(ins, callback)
+		
+		RunService.RenderStepped:Wait()
+		parent.Parent = nil
+		RunService.RenderStepped:Wait()
+		ins.Parent = nil
+		RunService.RenderStepped:Wait()
+		parent:Destroy()
+
+		local start = os.clock()
+		local timeout = 6
+
+		repeat
+			RunService.RenderStepped:Wait()
+			if os.clock() - start > timeout then
+				return
+			end
+		until done
+
+		error("Instance was incorrectly collected")
+	end)
+
+	it("should run tasks if parent is moved to nil then goes out of scope", function()
+		local done = false
+		local function callback()
+			done = true
+		end
+		do
+			local parent = Instance.new("Folder")
+			parent.Parent = workspace
+
+			local ins = Instance.new("Folder")
+			ins.Parent = Instance.new("Folder")
+
+			cleanupOnDestroy(ins, callback)
+			RunService.RenderStepped:Wait()
+
+			parent.Parent = nil
+		end
+
+		local start = os.clock()
+		local timeout = 3
+
+		repeat
+			RunService.RenderStepped:Wait()
+			if os.clock() - start > timeout then
+				error("Instance was held in memory for too long")
+			end
+		until done
 	end)
 end