"Successful" minimal polling version of ``cleanupOnDestroy_smart``. I'm really just putting this here so we can decide what to do.

**The good part:**
All the test cases now pass.

Speaking of test cases, I added 4 new test cases to directly examine this ancestry behavior and make sure there are no false positives. But I'm not sure I covered everything, and I can't think of more test cases. I'm looking for some other eyes to look over this (it's still in a draft phase), and help me come up with more test cases if necessary. 

I might also consider throwing the ``updateRefStrength()`` part of ``New.lua`` into ``cleanupTasks``.

**The bad part:**
This originally would crash benchmarking when throwing it under ``New.lua``. Dropping the benchmarking down to 1/10th of what it currently is lets the benchmarking run... But with the small issue that the game is literally unplayable as you get ~ 2 fps (I have a 2080 SUPER GPU)

Old benchmarking (cleanupOnDestroy):
```
[+] New
   [+] New with Children - array .................................................. 8.56 μs
   [+] New with Children - single ................................................. 9.90 μs
   [+] New with Children - state .................................................. 16.30 μs
   [+] New with OnChange .......................................................... 16.67 μs
   [+] New with OnEvent ........................................................... 14.53 μs
   [+] New with Parent - constant ................................................. 14.61 μs
   [+] New with Parent - state .................................................... 15.00 μs
   [+] New with properties - constant ............................................. 9.01 μs
   [+] New with properties - state ................................................ 16.72 μs
   [+] New without properties ..................................................... 5.13 μs
[+] Oklab
   [+] Convert from Oklab ......................................................... 0.27 μs
   [+] Convert to Oklab ........................................................... 0.35 μs
[+] Scheduler
   [+] Enqueue + run tasks - Single property, many instances, many priorities ..... 3.45 μs
   [+] Enqueue + run tasks - Single property, many instances, single priority ..... 3.55 μs
   [+] Enqueue + run tasks - Single property, single instance, single priority .... 1.23 μs
   [+] Enqueue callback to be called .............................................. 0.05 μs
   [+] Enqueue property change on instance ........................................ 0.06 μs
   [+] Run tasks - nothing queued ................................................. 0.03 μs
   [+] [1 call] Enqueue callback to be called ..................................... 0.30 μs
   [+] [1 call] Enqueue property change on instance ............................... 0.90 μs
[+] captureDependencies
   [+] Capture dependencies from callback ......................................... 0.13 μs
[+] updateAll
   [+] Update deep tree ........................................................... 2.06 μs
   [+] Update empty tree .......................................................... 0.33 μs
   [+] Update shallow tree ........................................................ 1.63 μs
   [+] Update tree with complex dependencies ...................................... 2.79 μs 
```

New benchmarking (cleanupOnDestroy_smart):
```
[+] New
   [+] New with Children - array .................................................. 7.67 μs
   [+] New with Children - single ................................................. 9.95 μs
   [+] New with Children - state .................................................. 14.92 μs
   [+] New with OnChange .......................................................... 17.76 μs
   [+] New with OnEvent ........................................................... 14.53 μs
   [+] New with Parent - constant ................................................. 14.54 μs
   [+] New with Parent - state .................................................... 15.99 μs
   [+] New with properties - constant ............................................. 8.78 μs
   [+] New with properties - state ................................................ 16.91 μs
   [+] New without properties ..................................................... 4.95 μs
[+] Oklab
   [+] Convert from Oklab ......................................................... 0.22 μs
   [+] Convert to Oklab ........................................................... 0.35 μs
[+] Scheduler
   [+] Enqueue + run tasks - Single property, many instances, many priorities ..... 3.32 μs
   [+] Enqueue + run tasks - Single property, many instances, single priority ..... 3.23 μs
   [+] Enqueue + run tasks - Single property, single instance, single priority .... 0.78 μs
   [+] Enqueue callback to be called .............................................. 0.04 μs
   [+] Enqueue property change on instance ........................................ 0.05 μs
   [+] Run tasks - nothing queued ................................................. 0.03 μs
   [+] [1 call] Enqueue callback to be called ..................................... 0.30 μs
   [+] [1 call] Enqueue property change on instance ............................... 0.60 μs
[+] captureDependencies
   [+] Capture dependencies from callback ......................................... 0.13 μs
[+] updateAll
   [+] Update deep tree ........................................................... 2.37 μs
   [+] Update empty tree .......................................................... 0.32 μs
   [+] Update shallow tree ........................................................ 2.01 μs
   [+] Update tree with complex dependencies ...................................... 3.04 μs
```